Centralize FontCustomPlatformData.h https://bugs.webkit.org/show_bug.cgi?id=251946
Reviewed by Brent Fulgham and Adrian Perez de Castro.

There were three different versions of this, cairo, win, and mac. When
removing the AppleWin build the cairo version of the file got included
so move the file to a single place and use guards around for each
variant.

Also rename the files according to conventions.

* Source/WebCore/Headers.cmake:
* Source/WebCore/PlatformMac.cmake:
* Source/WebCore/PlatformWin.cmake:
* Source/WebCore/PlatformWinCairo.cmake:
* Source/WebCore/SourcesCocoa.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/FontCustomPlatformData.h: Added.
* Source/WebCore/platform/graphics/cairo/FontCustomPlatformData.h: Removed.
* Source/WebCore/platform/graphics/freetype/FontCustomPlatformDataFreeType.cpp:
* Source/WebCore/platform/graphics/mac/FontCustomPlatformData.h: Removed.
* Source/WebCore/platform/graphics/mac/FontCustomPlatformDataMac.cpp: Renamed from Source\WebCore\platform\graphics\mac\FontCustomPlatformData.cpp.
* Source/WebCore/platform/graphics/win/FontCustomPlatformData.h: Removed.
* Source/WebCore/platform/graphics/win/FontCustomPlatformDataWin.cpp: Renamed from Source\WebCore\platform\graphics\win\FontCustomPlatformDataCairo.cpp.

Canonical link: https://commits.webkit.org/260041@main
