+++ Source/WebCore/PlatformWin.cmake
@@ -102,7 +102,6 @@ list(APPEND WebCore_PRIVATE_FRAMEWORK_HEADERS
     platform/graphics/opentype/FontMemoryResource.h
 
     platform/graphics/win/DIBPixelData.h
-    platform/graphics/win/FontCustomPlatformData.h
     platform/graphics/win/FullScreenController.h
     platform/graphics/win/FullScreenControllerClient.h
     platform/graphics/win/FullScreenWindow.h
